Fixed small error with text files
When retrieving the latest changes from github, the expected output files get reformatted so that every line ended with the regex [\r\n] although the files were only push with [\n] at the end of the lines. I simply changed the transactions to end with EOL which contains '\r\n" to avoid this problem in the future.